Fixed running Post Config hooks if docker or kubernetes configs are not available

+++ setup-container.sh
@@ -2,17 +2,25 @@
 PRE_CONFIG_HOOKS=/usr/local/lib/devcontainer/hooks.d/pre-start
 POST_CONFIG_HOOKS=/usr/local/lib/devcontainer/hooks.d/post-start
 
+# Pre Config hooks
 if [[ -d "${PRE_CONFIG_HOOKS}" ]] && [ "$(ls -A "${PRE_CONFIG_HOOKS}")" ]; then
     for HOOK in "${PRE_CONFIG_HOOKS}"/*; do
         ${HOOK}
     done
 fi
 
-jq 'del(.credsStore)' /home/vscode/.docker/hostconfig.json > /home/vscode/.docker/config.json
+# Setup Docker & Kubernetes configs
 sudo mkdir /root/.docker /root/.kube
-sudo ln -s /home/vscode/.docker/config.json /root/.docker/config.json
-cp /home/vscode/.kube/hostconfig /home/vscode/.kube/config
-sudo ln -s /home/vscode/.kube/config /root/.kube/config
+
+if [[ -e /home/vscode/.docker/hostconfig.json ]]; then
+    jq 'del(.credsStore)' /home/vscode/.docker/hostconfig.json > /home/vscode/.docker/config.json
+    sudo ln -s /home/vscode/.docker/config.json /root/.docker/config.json
+fi
+
+if [[ -e /home/vscode/.kube/hostconfig ]]; then
+    cp /home/vscode/.kube/hostconfig /home/vscode/.kube/config
+    sudo ln -s /home/vscode/.kube/config /root/.kube/config
+fi
 
 if [[ -e /home/vscode/.helm-repositories.yaml ]]; then
     mkdir -p /home/vscode/.config/helm 2>/dev/null || true
@@ -22,6 +30,7 @@ if [[ -e /home/vscode/.helm-repositories.yaml ]]; then
     sudo ln -sf /home/vscode/.helm-repositories.yaml /root/.config/helm/repositories.yaml
 fi
 
+# Post config hooks
 if [[ -d "${POST_CONFIG_HOOKS}" ]] && [ "$(ls -A "${POST_CONFIG_HOOKS}")" ]; then
     for HOOK in "${POST_CONFIG_HOOKS}"/*; do
         ${HOOK}
